# Conflicts:
#	acceptor.go

+++ internal/netpoll/defs_bsd_32bit.go
@@ -1,4 +1,4 @@
-//go:build amd64 || arm64 || ppc64 || ppc64le || mips64 || mips64le || riscv64
+//go:build (freebsd || dragonfly || darwin) && (386 || arm || mips || mipsle)
 
 /*
  * Copyright (c) 2023 Andy Pan.
@@ -19,4 +19,4 @@
 
 package netpoll
 
-type keventIdent = uint64
+type keventIdent = uint32
